requirements.txt
discord.py==2.0.1
beautifulsoup4==4.9.3
requests==2.25.1
